+++ engine/src/execution_graph/Context.cpp
@@ -10,11 +10,11 @@ Context::Context(const uint32_t token,
                  const std::string &logicalPlan,
                  const std::map<std::string, std::string>& config_options)
     : token_{token},
+      query_step{0},
+      query_substep{0},
       taskNodes_{taskNodes},
       masterNode_{masterNode},
       logicalPlan_{logicalPlan},
-      query_step{0},
-      query_substep{0},
       kernel_id_{0},
       config_options_{config_options} {}
 
@@ -37,7 +37,7 @@ std::vector<Node> Context::getAllOtherNodes(
   std::vector<Node> siblings(taskNodes_.size() - 1);
   size_t count = 0;
   for (size_t i = 0; i < taskNodes_.size(); i++) {
-    if (i != selfNodeIndex) {
+    if (i != static_cast<size_t>(selfNodeIndex)) {
       siblings[count] = taskNodes_[i];
       count++;
     }